数据分析活/电商订单数据分析/requirements.txt
# 数据处理
pandas==2.1.0
numpy==1.24.3

# 大数据组件
pyspark==3.5.0
apache-flink==1.17.0
kafka-python==2.0.2

# 数据库
mysql-connector-python==8.0.33
pymongo==4.5.0
redis==5.0.1

# 机器学习
scikit-learn==1.3.0
xgboost==2.0.0
prophet==1.1.4

# 可视化
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2
pyecharts==2.0.4

# Web服务
flask==3.0.0
requests==2.31.0

# 监控和日志
prometheus-client==0.17.1
python-logging==0.4.9.6

# 系统工具
psutil==5.9.5
docker==6.1.3
pyyaml==6.0.1

# 开发工具
jupyter==1.0.0
ipython==8.15.0 